+++ src/lib/SessionKeyPair.ts
@@ -1,5 +1,5 @@
 import { generateRandom64BitValue } from './crypto/_utils';
-import { generateECDHKeyPair } from './crypto/keys';
+import { generateECDHKeyPair } from './crypto/keys/generation';
 import { SessionKey } from './SessionKey';
 
 export class SessionKeyPair {
