+++ src/lib/bindings/gsc/ParcelCollection.spec.ts
@@ -6,7 +6,7 @@ import {
   generateStubCert,
   reSerializeCertificate,
 } from '../../_test_utils';
-import { generateRSAKeyPair } from '../../crypto/keys';
+import { generateRSAKeyPair } from '../../crypto/keys/generation';
 import { Certificate } from '../../crypto/x509/Certificate';
 import { InvalidMessageError } from '../../messages/InvalidMessageError';
 import { Parcel } from '../../messages/Parcel';
