+++ src/lib/bindings/gsc/PrivateNodeRegistrationRequest.ts
@@ -1,10 +1,10 @@
 import { ObjectIdentifier, OctetString, Primitive, verifySchema } from 'asn1js';
 
 import { makeHeterogeneousSequenceSchema, makeImplicitlyTaggedSequence } from '../../asn1';
-import { derDeserializeRSAPublicKey, derSerializePublicKey } from '../../crypto/keys';
 import { sign, verify } from '../../crypto/rsaSigning';
 import { InvalidMessageError } from '../../messages/InvalidMessageError';
 import { RELAYNET_OIDS } from '../../oids';
+import { derDeserializeRSAPublicKey, derSerializePublicKey } from '../../crypto/keys/serialisation';
 
 export class PrivateNodeRegistrationRequest {
   public static async deserialize(
