+++ src/lib/crypto/x509/Certificate.ts
@@ -4,8 +4,8 @@ import * as pkijs from 'pkijs';
 
 import * as oids from '../../oids';
 import { derDeserialize, generateRandom64BitValue } from '../_utils';
-import { getIdFromIdentityKey, getPublicKeyDigest } from '../keys';
-import { getEngineForPrivateKey } from '../webcrypto/engine';
+import { getIdFromIdentityKey, getPublicKeyDigest } from '../keys/generation';
+import { getEngineForPrivateKey } from '../pkijs';
 import { CertificateError } from './CertificateError';
 import FullCertificateIssuanceOptions from './FullCertificateIssuanceOptions';
 import { assertPkiType, assertUndefined } from '../cms/_utils';
