+++ src/lib/messages/payloads/CargoMessageSet.spec.ts
@@ -10,7 +10,7 @@ import {
   generateStubCert,
 } from '../../_test_utils';
 import { derDeserialize } from '../../crypto/_utils';
-import { generateRSAKeyPair } from '../../crypto/keys';
+import { generateRSAKeyPair } from '../../crypto/keys/generation';
 import { Certificate } from '../../crypto/x509/Certificate';
 import { CertificationPath } from '../../pki/CertificationPath';
 import { MAX_SDU_PLAINTEXT_LENGTH } from '../../ramf/serialization';
