+++ src/lib/nodes/Gateway.spec.ts
@@ -1,7 +1,7 @@
 import { addDays, setMilliseconds } from 'date-fns';
 
 import { reSerializeCertificate } from '../_test_utils';
-import { generateRSAKeyPair, getIdFromIdentityKey } from '../crypto/keys';
+import { generateRSAKeyPair, getIdFromIdentityKey } from '../crypto/keys/generation';
 import { Certificate } from '../crypto/x509/Certificate';
 import { MockKeyStoreSet } from '../keyStores/testMocks';
 import { CertificationPath } from '../pki/CertificationPath';
