+++ src/lib/nodes/PrivateGateway.spec.ts
@@ -3,11 +3,10 @@ import { addDays, setMilliseconds, subSeconds } from 'date-fns';
 import { arrayBufferFrom, reSerializeCertificate } from '../_test_utils';
 import { PrivateNodeRegistrationRequest } from '../bindings/gsc/PrivateNodeRegistrationRequest';
 import {
-  derSerializePublicKey,
   generateRSAKeyPair,
   getIdFromIdentityKey,
   getRSAPublicKeyFromPrivate,
-} from '../crypto/keys';
+} from '../crypto/keys/generation';
 import { Certificate } from '../crypto/x509/Certificate';
 import { MockKeyStoreSet } from '../keyStores/testMocks';
 import { CertificationPath } from '../pki/CertificationPath';
@@ -16,6 +15,7 @@ import { SessionKey } from '../SessionKey';
 import { SessionKeyPair } from '../SessionKeyPair';
 import { NodeError } from './errors';
 import { PrivateGateway } from './PrivateGateway';
+import { derSerializePublicKey } from '../crypto/keys/serialisation';
 
 const INTERNET_GATEWAY_INTERNET_ADDRESS = 'example.com';
