+++ src/lib/nodes/channels/GatewayChannel.spec.ts
@@ -9,7 +9,7 @@ import {
 } from '../../_test_utils';
 import { EnvelopedData, SessionEnvelopedData } from '../../crypto/cms/envelopedData';
 import { SignatureOptions } from '../../crypto/cms/SignatureOptions';
-import { generateRSAKeyPair, getIdFromIdentityKey } from '../../crypto/keys';
+import { generateRSAKeyPair, getIdFromIdentityKey } from '../../crypto/keys/generation';
 import { Certificate } from '../../crypto/x509/Certificate';
 import { MockKeyStoreSet } from '../../keyStores/testMocks';
 import { Cargo } from '../../messages/Cargo';
