+++ src/lib/nodes/channels/PrivateGatewayChannel.spec.ts
@@ -1,17 +1,17 @@
 import { addDays, setMilliseconds, subMinutes, subSeconds } from 'date-fns';
 
 import {
-  derSerializePublicKey,
   generateRSAKeyPair,
   getIdFromIdentityKey,
   getRSAPublicKeyFromPrivate,
-} from '../../crypto/keys';
+} from '../../crypto/keys/generation';
 import { Certificate } from '../../crypto/x509/Certificate';
 import { MockKeyStoreSet } from '../../keyStores/testMocks';
 import { CertificationPath } from '../../pki/CertificationPath';
 import { issueGatewayCertificate } from '../../pki/issuance';
 import { NodeCryptoOptions } from '../NodeCryptoOptions';
 import { PrivateGatewayChannel } from './PrivateGatewayChannel';
+import { derSerializePublicKey } from '../../crypto/keys/serialisation';
 
 let internetGatewayId: string;
 let internetGatewayPublicKey: CryptoKey;
