+++ src/lib/nodes/channels/PrivateGatewayChannel.ts
@@ -1,6 +1,6 @@
 import { addDays, subMinutes } from 'date-fns';
 
-import { getIdFromIdentityKey, getRSAPublicKeyFromPrivate } from '../../crypto/keys';
+import { getIdFromIdentityKey, getRSAPublicKeyFromPrivate } from '../../crypto/keys/generation';
 import { Certificate } from '../../crypto/x509/Certificate';
 import { CertificationPath } from '../../pki/CertificationPath';
 import { issueGatewayCertificate } from '../../pki/issuance';
