+++ src/lib/nodes/signatures/Signer.spec.ts
@@ -3,7 +3,7 @@ import { ObjectIdentifier, OctetString } from 'asn1js';
 import { arrayBufferFrom, generateStubCert } from '../../_test_utils';
 import { makeImplicitlyTaggedSequence } from '../../asn1';
 import { SignedData } from '../../crypto/cms/signedData';
-import { generateRSAKeyPair } from '../../crypto/keys';
+import { generateRSAKeyPair } from '../../crypto/keys/generation';
 import { Certificate } from '../../crypto/x509/Certificate';
 import { STUB_OID_VALUE, StubSigner } from './_test_utils';
